Migrate WithCustomerBaseCommand test to linux test suite for test coverage purposes
