+++ command_darwin_test.go
@@ -1,25 +1,11 @@
 package cmd
 
 import (
-	"os/exec"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func TestCommand_WithCustomBaseCommand(t *testing.T) {
-	cmd := NewCommand(
-		"echo $0",
-		WithCustomBaseCommand(exec.Command("/bin/bash", "-c")),
-	)
-
-	err := cmd.Execute()
-	assert.Nil(t, err)
-	// on darwin we use /bin/sh by default test if we're using bash
-	assert.NotEqual(t, "/bin/sh\n", cmd.Stdout())
-	assert.Equal(t, "/bin/bash\n", cmd.Stdout())
-}
 
 func TestCommand_ExecuteStderr(t *testing.T) {
 	cmd := NewCommand(">&2 echo hello")
